# static/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "FUB to GHL service",
    "description": "API for managing leads, updating tags, and handling tasks.",
    "version": "1.0.0"
  },
  "host": "3.98.147.51:5007",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "ApiKeyAuth": {
      "type": "apiKey",
      "in": "header",
      "name": "X-API-KEY"
    }
  },
  "paths": {
    "/lead": {
      "post": {
        "summary": "Create a new lead or add a new inquiry note",
        "description": "Creates a lead in the GHL system. If it already exists, adds an inquiry note",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "person": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer"
                    },
                    "firstName": {
                      "type": "string"
                    },
                    "lastName": {
                      "type": "string"
                    },
                    "addresses": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "city": {
                            "type": "string"
                          },
                          "state": {
                            "type": "string"
                          }
                        }
                      }
                    },
                    "customMLSNumber": {
                      "type": "string"
                    },
                    "customListingType": {
                      "type": "string"
                    },
                    "customAddress": {
                      "type": "string"
                    },
                    "customCity": {
                      "type": "string"
                    },
                    "customProvince": {
                      "type": "string"
                    },
                    "customFB4SLeadID": {
                      "type": "string"
                    },
                    "customFB4SRCAURL": {
                      "type": "string"
                    },
                    "customFB4SInquiriesCounter": {
                      "type": "string"
                    },
                    "customChromeExtensionLink": {
                      "type": "string"
                    },
                    "customListingURLPath": {
                      "type": "string"
                    },
                    "customExpectedPriceRange": {
                      "type": "string"
                    },
                    "customBuyerProfileFB4S": {
                      "type": "string"
                    },
                    "customAssignedNotFromWillowAt": {
                      "type": "string"
                    },
                    "customStage": {
                      "type": "string"
                    },
                    "customPrice": {
                      "type": "string"
                    },
                    "customClosingAnniversary": {
                      "type": "string"
                    },
                    "customYlopoSellerReport": {
                      "type": "string"
                    },
                    "customWhoareyou": {
                      "type": "string"
                    },
                    "customLastActivity": {
                      "type": "string"
                    },
                    "customCloseDate": {
                      "type": "string"
                    },
                    "customYlopoStarsLink": {
                      "type": "string"
                    },
                    "customOldID": {
                      "type": "string"
                    },
                    "emails": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "value": {
                            "type": "string"
                          }
                        }
                      }
                    },
                    "phones": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "value": {
                            "type": "string"
                          }
                        }
                      }
                    },
                    "tags": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "selected_realtor_email": {
                      "type": "string"
                    }
                  }
                },
                "property": {
                  "type": "object",
                  "properties": {
                    "street": {
                      "type": "string"
                    },
                    "city": {
                      "type": "string"
                    },
                    "state": {
                      "type": "string"
                    },
                    "code": {
                      "type": "string"
                    },
                    "mlsNumber": {
                      "type": "string"
                    },
                    "url": {
                      "type": "string"
                    }
                  }
                },
                "type": {
                  "type": "string"
                },
                "source": {
                  "type": "string"
                },
                "system": {
                  "type": "string"
                },
                "description": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                }
              },
              "required": [
                "person"
              ]
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Lead successfully created",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Lead successfully created"
                },
                "contact": {
                  "type": "object",
                  "properties": {
                    "city": {
                      "type": "string",
                      "example": "N/A"
                    },
                    "country": {
                      "type": "string",
                      "example": "CA"
                    },
                    "customField": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "fieldValue": {
                            "type": "string"
                          },
                          "id": {
                            "type": "string"
                          }
                        }
                      },
                      "example": [
                        {
                          "fieldValue": "Laundromats",
                          "id": "R3CCQhYeG4kZ5NSTW5vk"
                        },
                        {
                          "fieldValue": "https://www.findbusinesses4sale.com/rca/eyJjaGF0X2lkIjoxODY0NSwiY29ubmVjdGlvbl9vd25lcl9pZCI6IjE4NzA1In0=/",
                          "id": "F4Bkzj3AXKtBiri6S3Xe"
                        }
                      ]
                    },
                    "dateAdded": {
                      "type": "string",
                      "format": "date-time",
                      "example": "2025-03-28T14:45:09.882Z"
                    },
                    "email": {
                      "type": "string",
                      "example": "mconcha@burmanu.ca"
                    },
                    "emailLowerCase": {
                      "type": "string",
                      "example": "mconcha@burmanu.ca"
                    },
                    "firstName": {
                      "type": "string",
                      "example": "Sara"
                    },
                    "firstNameLowerCase": {
                      "type": "string",
                      "example": "sara"
                    },
                    "fullNameLowerCase": {
                      "type": "string",
                      "example": "sara olit"
                    },
                    "id": {
                      "type": "string",
                      "example": "TBfGmIEsBVStLeBGAwiG"
                    },
                    "lastName": {
                      "type": "string",
                      "example": "Olit"
                    },
                    "lastNameLowerCase": {
                      "type": "string",
                      "example": "olit"
                    },
                    "locationId": {
                      "type": "string",
                      "example": "iCIim8m7X3EQyFY9rQ7H"
                    },
                    "phone": {
                      "type": "string",
                      "example": "+17789800752"
                    },
                    "source": {
                      "type": "string",
                      "example": "FB4S"
                    },
                    "state": {
                      "type": "string",
                      "example": "N/A"
                    },
                    "tags": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      },
                      "example": [
                        "fb4s"
                      ]
                    },
                    "type": {
                      "type": "string",
                      "example": "lead"
                    }
                  }
                }
              }
            }
          },
          "200": {
            "description": "User was not created, but an inquiry note was added",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "User was not created, already exists. Added inquiry note if provided"
                },
                "contact": {
                  "type": [
                    "string",
                    "null"
                  ],
                  "example": null
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized request",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Unauthorized"
                }
              }
            }
          },
          "406": {
            "description": "Validation error in the request payload",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Validation error {'person': ['Missing data for required field.']}"
                },
                "contact": {
                  "type": [
                    "string",
                    "null"
                  ],
                  "example": null
                }
              }
            }
          },
          "400": {
            "description": "General error occurred while creating lead",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "error: Unexpected issue occurred"
                },
                "contact": {
                  "type": [
                    "string",
                    "null"
                  ],
                  "example": null
                }
              }
            }
          }
        }
      }
    },
...
    "/lead/{lead_id}/followers": {
      "post": {
        "summary": "add followers(collaborators) to lead",
        "description": "add followers(collaborators) to lead",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "followers": {
                      "type": "array",
                      "items": {
                        "type": "string",
                        "example": "Fo7TmuqqwHIM22eno39t"
                      }
                    }
              },
              "required": [
                "followers"
              ]
            }
          }
        ],
        "responses": {
          "201": {
            "description": "follower added successfully",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "followers added successfully"
                }
              }
            }
          },
          "202": {
            "description": "Lead was not found by this email",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Something went wrong, followers weren't added"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized request",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Unauthorized"
                }
              }
            }
          },
          "406": {
            "description": "Validation error in the request payload",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Validation error {'followers': ['Missing data for required field.']}"
                }
              }
            }
          },
          "400": {
            "description": "General error occurred while adding followers",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Error: Unexpected issue occurred"
                }
              }
            }
          }
        }
      }
    },
...
    "/users": {
      "get": {
        "summary": "Get users(team members) list",
        "description": "Retrieves users(team members) list",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved users list",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Successfully get users list"
                },
                "users": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/User"
                  }
                }
              }
            }
          },
          "202": {
            "description": "User was not found by this email",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "There is no users currently in this location"
                },
                "users": {
                  "type": [
                    "string",
                    "null"
                  ],
                  "example": null
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized request",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Unauthorized"
                }
              }
            }
          },
          "400": {
            "description": "General error occurred while getting lead by id",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Error while getting a users list 'error'"
                },
                "users": {
                  "type": [
                    "string",
                    "null"
                  ],
                  "example": null
                }
              }
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "User": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "firstName": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "phone": {
          "type": "string"
        },
        "permissions": {
          "type": "object",
          "additionalProperties": {
            "type": "boolean"
          }
        },
        "roles": {
          "type": "object",
          "properties": {
            "locationIds": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "role": {
              "type": "string"
            },
            "type": {
              "type": "string"
            }
          }
        }
      }
    }
  }
}